+++ test/functional/p2p_pos_doublespend.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
+# Copyright (c) 2019 The PIVX Core developers
+# Distributed under the MIT software license, see the accompanying
+# file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
 '''
 Covers the scenario of a valid PoS block with a valid coinstake transaction where the
 coinstake input prevout is double spent in one of the other transactions in the same block.
@@ -7,10 +10,10 @@ coinstake input prevout is double spent in one of the other transactions in the 
 
 from time import sleep
 
-from base_test import PIVX_FakeStakeTest
+from fake_stake.base_test import PIVX_FakeStakeTest
 
 
-class Test_05(PIVX_FakeStakeTest):
+class PoSDoubleSpend(PIVX_FakeStakeTest):
 
     def run_test(self):
         self.description = "Covers the scenario of a valid PoS block with a valid coinstake transaction where the coinstake input prevout is double spent in one of the other transactions in the same block."
@@ -46,3 +49,6 @@ class Test_05(PIVX_FakeStakeTest):
             raise AssertionError("TEST FAILED")
 
         self.log.info("%s PASSED" % self.__class__.__name__)
+
+if __name__ == '__main__':
+    PoSDoubleSpend().main()
