+++ test/functional/p2p_pos_fakestake_accepted.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
+# Copyright (c) 2019 The PIVX Core developers
+# Distributed under the MIT software license, see the accompanying
+# file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
 '''
 Covers the scenario of a valid PoS block where the coinstake input prevout is spent on main chain,
 but not on the fork branch. These blocks must be accepted.
@@ -7,9 +10,9 @@ but not on the fork branch. These blocks must be accepted.
 
 from time import sleep
 
-from base_test import PIVX_FakeStakeTest
+from fake_stake.base_test import PIVX_FakeStakeTest
 
-class Test_02(PIVX_FakeStakeTest):
+class PoSFakeStakeAccepted(PIVX_FakeStakeTest):
 
     def run_test(self):
         self.description = "Covers the scenario of a valid PoS block where the coinstake input prevout is spent on main chain, but not on the fork branch. These blocks must be accepted."
@@ -52,3 +55,6 @@ class Test_02(PIVX_FakeStakeTest):
             raise AssertionError("TEST FAILED")
 
         self.log.info("%s PASSED" % self.__class__.__name__)
+
+if __name__ == '__main__':
+    PoSFakeStakeAccepted().main()
