+++ test/functional/p2p_zpos_fakestake_accepted.py
@@ -1,26 +1,37 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-'''
-Covers the scenario of a zPoS block where the coinstake input is a zerocoin spend
-of an already spent coin.
-'''
+# Copyright (c) 2019 The PIVX Core developers
+# Distributed under the MIT software license, see the accompanying
+# file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-from time import sleep
+'''
+Performs the same check as in Test_02 verifying that zPoS forked blocks that stake a zerocoin which is spent on mainchain on an higher block are still accepted.
+'''
 
 from test_framework.authproxy import JSONRPCException
+from fake_stake.base_test import PIVX_FakeStakeTest
+from time import sleep
 
-from base_test import PIVX_FakeStakeTest
+class zPoSFakeStakeAccepted(PIVX_FakeStakeTest):
 
-class Test_03(PIVX_FakeStakeTest):
+
+    def set_test_params(self):
+        ''' Setup test environment
+        :param:
+        :return:
+        '''
+        self.setup_clean_chain = True
+        self.num_nodes = 1
+        self.extra_args = [['-staking=1', '-debug=net', '-zpivstake']] * self.num_nodes
 
     def run_test(self):
-        self.description = "Covers the scenario of a zPoS block where the coinstake input is a zerocoin spend of an already spent coin."
+        self.description = "Performs the same check as in Test_02 verifying that zPoS forked blocks that stake a zerocoin which is spent on mainchain on an higher block are still accepted."
         self.init_test()
 
-        DENOM_TO_USE = 5000         # zc denomination
-        INITAL_MINED_BLOCKS = 321   # First mined blocks (rewards collected to mint)
-        MORE_MINED_BLOCKS = 301     # More blocks mined before spending zerocoins
-        self.NUM_BLOCKS = 2         # Number of spammed blocks
+        DENOM_TO_USE = 1000  # zc denomination
+        INITAL_MINED_BLOCKS = 321
+        MORE_MINED_BLOCKS = 301
+        FORK_DEPTH = 75
+        self.NUM_BLOCKS = 2
 
         # 1) Starting mining blocks
         self.log.info("Mining %d blocks to get to zPOS activation...." % INITAL_MINED_BLOCKS)
@@ -44,7 +55,7 @@ class Test_03(PIVX_FakeStakeTest):
 
             if initial_mints % 5 == 0:
                 self.log.info("Minted %d coins" % initial_mints)
-            if initial_mints >= 70:
+            if initial_mints >= 20:
                 break
             balance = self.node.getbalance("*", 100)
         self.log.info("Minted %d coins in the %d-denom, remaining balance %d", initial_mints, DENOM_TO_USE, balance)
@@ -55,15 +66,20 @@ class Test_03(PIVX_FakeStakeTest):
         self.node.generate(MORE_MINED_BLOCKS)
         sleep(2)
         mints = self.node.listmintedzerocoins(True, True)
+        sleep(1)
         mints_hashes = [x["serial hash"] for x in mints]
 
         # This mints are not ready spendable, only few of them.
         self.log.info("Got %d confirmed mints" % len(mints_hashes))
 
-        # 4) spend mints
+        # 4) Start mining again so that spends get confirmed in a block.
+        self.log.info("Mining 200 more blocks...")
+        self.node.generate(200)
+        sleep(2)
+
+        # 5) spend mints
         self.log.info("Spending mints in block %d..." % self.node.getblockcount())
         spends = 0
-        spent_mints = []
         for mint in mints_hashes:
             # create a single element list to pass to RPC spendzerocoinmints
             mint_arg = []
@@ -72,29 +88,30 @@ class Test_03(PIVX_FakeStakeTest):
                 self.node.spendzerocoinmints(mint_arg)
                 sleep(1)
                 spends += 1
-                spent_mints.append(mint)
             except JSONRPCException as e:
                 self.log.warning(str(e))
                 continue
         sleep(1)
         self.log.info("Successfully spent %d mints" % spends)
 
-        # 5) Start mining again so that spends get confirmed in a block.
-        self.log.info("Mining 5 more blocks...")
-        self.node.generate(5)
+        self.log.info("Mining 6 more blocks...")
+        self.node.generate(6)
         sleep(2)
 
         # 6) Collect some prevouts for random txes
         self.log.info("Collecting inputs for txes...")
-        spending_utxo_list = self.node.listunspent()
+        utxo_list = self.node.listunspent()
         sleep(1)
 
-        # 7) Create "Fake Stake" blocks and send them
-        self.log.info("Creating Fake stake zPoS blocks...")
-        err_msgs = self.test_spam("Main", mints, spending_utxo_list=spending_utxo_list, fZPoS=True)
+        # 7) Create valid forked zPoS blocks and send them
+        self.log.info("Creating stake zPoS blocks...")
+        err_msgs = self.test_spam("Fork", mints, spending_utxo_list=utxo_list, fZPoS=True, fRandomHeight=True, randomRange=FORK_DEPTH, randomRange2=50, fMustPass=True)
 
         if not len(err_msgs) == 0:
             self.log.error("result: " + " | ".join(err_msgs))
             raise AssertionError("TEST FAILED")
 
         self.log.info("%s PASSED" % self.__class__.__name__)
+
+if __name__ == '__main__':
+    zPoSFakeStakeAccepted().main()
